Experimenting with package layout to make import paths cleaner
Right now we publish packages from the root directory and include the
lib directory in the package. This means our imports path becomes

    `@splunk/otel/lib/` instead of `@splunk/otel/`

This is somewhat displeasing. It would be nicer if users could import
from `@splunk/otel` and be able to auto instrument with `node -r
@splunk/otel/instrument` instead of @splunk/otel/lib/instrument`.

This commit copies the package.json into the lib directory as part of
the publish command, CDs into the lib directory and publishes the
package from there.